Lots of refactoring + make unit tests work again!
